requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asn1crypto==1.5.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==42.0.8
distro==1.9.0
filelock==3.15.4
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
openai==1.41.0
packaging==24.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.12.0
sortedcontainers==2.4.0
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
